% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleCas12a6plexgRNALibraryOligo.R
\name{assembleCas12a6plexgRNALibraryOligo}
\alias{assembleCas12a6plexgRNALibraryOligo}
\title{Function to assemble 6-plex oligo for pooled synthesis for Cas12a 6-plex gRNA expression
The oligo will be PCR'd using adaptor sequences, then digested with BsmbI and ligated into backbone.
by Chris Hsiung 2022-02-28}
\usage{
assembleCas12a6plexgRNALibraryOligo(
  pos1name,
  pos1spacer,
  pos2name,
  pos2spacer,
  pos3name,
  pos3spacer,
  pos4name,
  pos4spacer,
  pos5name,
  pos5spacer,
  pos6name,
  pos6spacer,
  Adaptor5p,
  Adaptor3p,
  DR1 = "AATTTCTACTGTCGTAGAT",
  DR16 = "AATTCCTACTATTGTAGGT",
  DR18 = "AATTCCTACTCTAGTAGGT",
  DR10 = "AATTCCTACTCTCGTAGGT",
  DR3 = "AATTTCTACTCTAGTAGAT",
  BsmbI5p = "cgtctcAAGAT",
  BsmbI3p = "AATTcgagacg",
  outputdir = getwd()
)
}
\value{

}
\description{
Function to assemble 6-plex oligo for pooled synthesis for Cas12a 6-plex gRNA expression
The oligo will be PCR'd using adaptor sequences, then digested with BsmbI and ligated into backbone.
by Chris Hsiung 2022-02-28
}
