% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleCas9DualgRNALibraryOligo.R
\name{assembleCas9DualgRNALibraryOligo}
\alias{assembleCas9DualgRNALibraryOligo}
\title{Function to facilitate assembly of oligo pool synthesis for Cas9 dual gRNA expression as in Replogle et al 2020
BsmBIstuffer defaults to sequence for containing BsmBI sitesß
REseq5p defaults to contain BstxI site
REseq3p defaults to contain BlpI site}
\usage{
assembleCas9DualgRNALibraryOligo(
  spacer1,
  spacer2,
  BsmBIstuffer = "gtttcagagcgagacgtgcctgcaggatacgtctcagaaacatg",
  REseq5p = "ccaccttgttg",
  REseq3p = "gtttaagagctaagctg",
  Adapter5p,
  Adapter3p
)
}
\value{
character containing assembled oligo sequence
}
\description{
Function to facilitate assembly of oligo pool synthesis for Cas9 dual gRNA expression as in Replogle et al 2020
BsmBIstuffer defaults to sequence for containing BsmBI sitesß
REseq5p defaults to contain BstxI site
REseq3p defaults to contain BlpI site
}
