% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_quantilethresperc.R
\name{calculate_quantilethresperc}
\alias{calculate_quantilethresperc}
\title{Calculate Quantile Threshold Percentages And Counts}
\usage{
calculate_quantilethresperc(
  data,
  group_var,
  thres_var = "Spacer",
  thresname_var = "NT-3",
  value_var,
  quantilethres_var = "0.05",
  round_var = 1
)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{group_var}{A character vector specifying the grouping variable(s).}

\item{thres_var}{A character vector specifying the threshold variable (default is 'Spacer').}

\item{thresname_var}{A character vector specifying the threshold name variable (default is 'NT-3').}

\item{value_var}{A character vector specifying the value variable.}

\item{quantilethres_var}{A character vector specifying the quantile threshold variable (default is '0.05').}

\item{round_var}{An integer specifying the number of decimal places to round the percentages (default is 1).}
}
\description{
This function calculates the quantile threshold percentages based on the input data.
Warning: roxygen documentation generated by chatgpt
}
