% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_name.R
\name{clean_name}
\alias{clean_name}
\title{Clean Name}
\usage{
clean_name(df, strtoremove)
}
\arguments{
\item{df}{A data frame.}

\item{strtoremove}{A character or pattern to be removed from column names.}
}
\value{
A data frame with modified column names.
}
\description{
Remove specified characters or patterns from column names of a data frame.
}
\examples{
df <- data.frame(First.Name = c("John", "Jane"),
Last.Name = c("Doe", "Smith"))
clean_name(df, "\.")
# Output:
# FirstName LastName
# 1 John Doe
# 2 Jane Smith

}
