% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_to_global.R
\name{copy_to_global}
\alias{copy_to_global}
\title{Copy Variables from One Environment to the Global Environment}
\usage{
copy_to_global(source_env)
}
\arguments{
\item{source_env}{An environment from which to copy variables.
This should be a valid R environment.}
}
\value{
Invisible NULL. The function is called for its side effect
of copying variables to the global environment.
}
\description{
This function copies all variables from a specified source environment
to the global environment.
}
\examples{
env_example <- new.env()
env_example$a <- 1
env_example$b <- 2
copy_to_global(env_example)
print(a)  # Should print 1
print(b)  # Should print 2
}
