% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countBlpI.R
\name{countBlpI}
\alias{countBlpI}
\title{sum total occurrence of sites for a dsDNA sequence (input one strand)}
\usage{
countBlpI(x)
}
\value{
numerical count of the number of sites found
}
\description{
sum total occurrence of sites for a dsDNA sequence (input one strand)
}
