% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta2df.R
\name{fasta2df}
\alias{fasta2df}
\title{wrapper for seqinr::read.fasta}
\usage{
fasta2df(fastafile)
}
\value{
Returns data frame with columns 'ID' and 'sequence'
}
\description{
wrapper for seqinr::read.fasta
}
