% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta2targetsites.R
\name{fasta2targetsites}
\alias{fasta2targetsites}
\title{By Chris Hsiung 2022-02-15}
\usage{
fasta2targetsites(
  fastafile,
  pam = "TTTV",
  altpams = c("TTYN", "CTTV", "RTTC", "TATM", "CTCC", "TCCC", "TACA", "TTTT")
)
}
\value{
data frame with target sites
}
\description{
By Chris Hsiung 2022-02-15
}
