% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowjo2df.R
\name{flowjo2df}
\alias{flowjo2df}
\title{Function to read csv file exported from flowjo. The original .fcs files from which the .csv file was generated must contain keywords. Parses sampleID from $FIL keyword and platename from $PLATENAME keyword and date from $DATE keyword.
Returns data frame
Intended for use within the function processFlowjoExportDir() for processing an entire directory of .csv files.
Updated 2022-08-02 to remove additional header lines inserted by Arc Attune that do not start with '$'
Updated 2023-04-04 added additional regex matches for identifying datacsv in case some variables are not exported from flowjo
Updated 2023-06-12 to allow IDvar as input
Warning: roxygen documentation generated by chatgpt}
\usage{
flowjo2df(file, IDvar)
}
\arguments{
\item{file}{A character string specifying the path of the flow cytometry file to be converted.}

\item{IDvar}{A character string specifying the name of the ID variable to be added to the dataframe.}
}
\value{
A dataframe with the flow cytometry data. In the case where the file contains no data, the function returns NULL.
}
\description{
#' Convert a flow cytometry file to a dataframe
}
\details{
This function takes a flow cytometry file and transforms it into a dataframe.
It also extracts meta information such as plate name, file name and date from the file.
The function also allows you to add an ID variable.
}
\examples{
\dontrun{
flowjo2df(file = "path_to_file.fcs", IDvar = "sample_ID")
}
}
