% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestCRISPickSpacers.R
\name{getBestCRISPickSpacers}
\alias{getBestCRISPickSpacers}
\title{Get Best CRISPick Spacers. As written is currently intended for CRIPRko output from CRISPick.}
\usage{
getBestCRISPickSpacers(
  CRISPickfile,
  TTTVonly = TRUE,
  checkBsmbI = TRUE,
  removeTTTT = TRUE
)
}
\arguments{
\item{CRISPickfile}{A data frame containing CRISPick results. Minimally must contain 'sgRNA Sequence' column, which must consist of DNA bases only}

\item{TTTVonly}{logical indicates whether to keep only TTTV PAM spacers; defaults to TRUE}

\item{checkBsmbI}{logical indicates whether to remove spacers containing BsmbI sites; defaults to TRUE}

\item{removeTTTT}{logical indicates whether to remove spacers containing TTTT or start with TTT; defaults to TRUE}
}
\value{
After filtering out TTTT, TTTstart, BsmbI sites (optional), Offtarget matches to 'MAX',  returns a data frame sorted by On-Target Efficacy Score.
}
\description{
Get Best CRISPick Spacers. As written is currently intended for CRIPRko output from CRISPick.
}
\author{
Chris Hsiung
}
