% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBothStrandsFASTA.R
\name{getBothStrandsFASTA}
\alias{getBothStrandsFASTA}
\title{by Chris Hsiung, updated 2022-02-15}
\usage{
getBothStrandsFASTA(
  fastafile,
  writefasta = TRUE,
  outputfile = paste0(fastafile, "_bothstrands")
)
}
\value{
Return data frame containing both strands. Also writes a FASTA file by default.
}
\description{
by Chris Hsiung, updated 2022-02-15
}
