% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCas9gRNAoligos_BstxIBlpI.R
\name{getCas9gRNAoligos_BstxIBlpI}
\alias{getCas9gRNAoligos_BstxIBlpI}
\title{for designing oligos to be cloned into BstxI/BlpI site (e.g. pLG1 vectors) for Cas9 gRNA
returns data frame of oligos}
\usage{
getCas9gRNAoligos_BstxIBlpI(name, spacer, spacerlength = 20)
}
\description{
for designing oligos to be cloned into BstxI/BlpI site (e.g. pLG1 vectors) for Cas9 gRNA
returns data frame of oligos
}
