% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiffEnrichment.R
\name{getDiffEnrichment}
\alias{getDiffEnrichment}
\title{This function calculates differential enrichment from screen raw read counts}
\usage{
getDiffEnrichment(
  df,
  initcount,
  finalcount,
  spacertype = "spacertype",
  negctrllabel = "negctrl"
)
}
\arguments{
\item{df}{is a data frame, must contain column names matching initcount, finalcount, and spacertype}

\item{initcount}{is a string for name of column in df containing raw read counts of initial condition (e.g. T0 in growth screen)}

\item{finalcount}{is a string for name of a column in df containing raw read counts of final condition (e.g. Tfinal in growth screen)}
}
\value{
a data frame with diffenrich added as column. This can be aggregated (e.g. take average across replicates) before being transformed (e.g. log2).
}
\description{
This function calculates differential enrichment from screen raw read counts
}
