% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethammingdist.R
\name{gethammingdist}
\alias{gethammingdist}
\title{function for getting the hamming distance between 2 strings}
\usage{
gethammingdist(str1, str2)
}
\description{
function for getting the hamming distance between 2 strings
}
