% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHumanBiomaRt.R
\name{getHumanBiomaRt}
\alias{getHumanBiomaRt}
\title{wrapper for getting ensemblID association with gene symbols
#'@export}
\usage{
getHumanBiomaRt()
}
\description{
wrapper for getting ensemblID association with gene symbols
#'@export
}
