% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLibraryOligoCas12a6plexgRNA.R
\name{getLibraryOligoCas12a6plexgRNA}
\alias{getLibraryOligoCas12a6plexgRNA}
\title{Function to assemble oligo for pooled synthesis for Cas12a 6-plex gRNA expression. The oligo will consist of: Adaptor5p-BsmbI5p-pos1spacer-DR1-pos2spacer-DR16-pos3spacer-DR18-pos4spacer-DR10-pos5spacer-DR3-pos6spacer-BsmbI3p-Adaptor3p.
DR sequences are taken from Deweirdt et al., Nat. Biotech 2020.
Assumes the oligo will be PCR'd using user-provided adaptor sequences, then digested with BsmbI and ligated into backbone.  The BsmbI5p and BsmbI3p sites are hard-coded to be compatible with vector backbone containing 5' DR and 3' DR sequences, which are not encoded in the oligo itself (e.g. compatible with vectors designs of pRG212, pCH39, pCH49, etc.).}
\usage{
getLibraryOligoCas12a6plexgRNA(
  pos1name,
  pos1spacer,
  pos2name,
  pos2spacer,
  pos3name,
  pos3spacer,
  pos4name,
  pos4spacer,
  pos5name,
  pos5spacer,
  pos6name,
  pos6spacer,
  Adaptor5p,
  Adaptor3p,
  sep = "_",
  outputdir = getwd()
)
}
\arguments{
\item{pos1name, pos2name, pos3name, pos4name, pos5name, pos6name}{string containing name for each spacer.}

\item{pos1spacer, pos2spacer, pos3spacer, pos4spacer, pos5spacer, pos6spacer}{string containing only spacer DNA sequences of length 19-23nt. Must not contain BsmbI sites. Must not contain TTTT. Must not start with TTT, because the last base of each DR is a T and that would form an undesirable TTTT at the junction.}

\item{Adaptor5p, Adaptor3p}{string containing PCR adaptor sequences to be used for subpool amplification (recommend taking from https://weissman.wi.mit.edu/crispr/)
The final oligo is checked for undesirable BsmbI sites (in excess of the 2 that should be there); if fails, throws error.}
}
\value{
A list containing oligoname and oligosequence. oligoname is in the format 'pos1name_pos2name_pos3name_pos4name_pos5name_pos6name'.
}
\description{
Function to assemble oligo for pooled synthesis for Cas12a 6-plex gRNA expression. The oligo will consist of: Adaptor5p-BsmbI5p-pos1spacer-DR1-pos2spacer-DR16-pos3spacer-DR18-pos4spacer-DR10-pos5spacer-DR3-pos6spacer-BsmbI3p-Adaptor3p.
DR sequences are taken from Deweirdt et al., Nat. Biotech 2020.
Assumes the oligo will be PCR'd using user-provided adaptor sequences, then digested with BsmbI and ligated into backbone.  The BsmbI5p and BsmbI3p sites are hard-coded to be compatible with vector backbone containing 5' DR and 3' DR sequences, which are not encoded in the oligo itself (e.g. compatible with vectors designs of pRG212, pCH39, pCH49, etc.).
}
\author{
Chris Hsiung
}
