% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLibraryoligoCas12aDoublegRNA.R
\name{getLibraryOligoCas12aDoublegRNA}
\alias{getLibraryOligoCas12aDoublegRNA}
\title{Function to assemble oligo for pooled synthesis for Cas12a dual gRNA expression. The oligo will consist of: Adaptor5p-BsmbI5p-pos1spacer-DR1-pos2spacer-BsmbI3p-Adaptor3p. DR sequences are taken from Deweirdt et al., Nat. Biotech 2020.
Assumes the oligo will be PCR'd using user-provided adaptor sequences, then digested with BsmbI and ligated into backbone.  The BsmbI5p and BsmbI3p sites are hard-coded to be compatible with vector backbone containing 5' DR and 3' DR sequences, which are not encoded in the oligo itself (e.g. compatible with vectors designs of pRG212, pCH39, pCH49, etc.).}
\usage{
getLibraryOligoCas12aDoublegRNA(
  pos1name = "None",
  pos1spacer,
  pos2name = "None",
  pos2spacer,
  Adaptor5p,
  Adaptor3p,
  sep = "_"
)
}
\arguments{
\item{pos1name, pos2name}{String containing name for each spacer. Defaults to "None".}

\item{pos1spacer, pos2spacer}{string containing only spacer DNA sequences of length 19-23nt. Must not contain BsmbI sites. Must not contain TTTT. Must not start with TTT, because the last base of each DR is a T and that would form an undesirable TTTT at the junction.}

\item{Adaptor5p, Adaptor3p}{string containing PCR adaptor sequences to be used for subpool amplification (recommend taking from https://weissman.wi.mit.edu/crispr/)
The final oligo is checked for undesirable BsmbI sites (in excess of the 2 that should be there); if fails, throws error.}

\item{sep}{String that will be used to separate names. Defaults to '_'.}
}
\value{
A list containing oligoname and oligosequence.
}
\description{
Function to assemble oligo for pooled synthesis for Cas12a dual gRNA expression. The oligo will consist of: Adaptor5p-BsmbI5p-pos1spacer-DR1-pos2spacer-BsmbI3p-Adaptor3p. DR sequences are taken from Deweirdt et al., Nat. Biotech 2020.
Assumes the oligo will be PCR'd using user-provided adaptor sequences, then digested with BsmbI and ligated into backbone.  The BsmbI5p and BsmbI3p sites are hard-coded to be compatible with vector backbone containing 5' DR and 3' DR sequences, which are not encoded in the oligo itself (e.g. compatible with vectors designs of pRG212, pCH39, pCH49, etc.).
}
\examples{
getLibraryOligoCas12aDoublegRNA( pos1name = 'CD81-1', pos1spacer = 'GGCGCGACCCCCAGGAAGGTCTC', pos2name = 'B2M-1', pos2spacer = 'ATATAAGTGGAGGCGTCGCGCTG', Adaptor5p = 'ATTTTGCCCCTGGTTCTT', Adaptor3p = 'CCAGTTCATTTCTTAGGG')
}
\author{
Chris Hsiung
}
