% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOligosCas12a5plexgRNAVariantDR_BsmbI.R
\name{getOligosCas12a5plexgRNAVariantDR_BsmbI}
\alias{getOligosCas12a5plexgRNAVariantDR_BsmbI}
\title{for individually cloning Cas12a 5-plex spacer array into vector with BsmBI site (e.g. pRG212). Modified from protocol by Qingzhou Chen and Junwei Shi. Variant DR sequences from Deweirdt et al. are used to minimize recombination. Also as much as possible avoid TTTCT motif (weak Pol III terminator)
Returns a list with two data frames, one containing oligos/ultramers, the other containing gene block. User can choose to use either one for synthesis, but gene block is more cost effective.
For gene block approach, by default insert is appended with homology regions (Gibson5p and Gibson3p as input arguments) compatible with Gibson assembly into pCH49.
For oligos/ultramer approach: nomenclature for naming sense and antisense strands is as follows: sense1 anneals with antisense1, sense2 anneals with antisense2, sense 3 anneals with antisense3, etc. These prefixes are printed at the front of the string to facilitate setting up annealing reactions because the IDT tubes truncates the end of the long string.
by Chris Hsiung, updated 2021-12-23}
\usage{
getOligosCas12a5plexgRNAVariantDR_BsmbI(
  pos1name,
  pos1spacer,
  pos2name,
  pos2spacer,
  pos3name,
  pos3spacer,
  pos4name,
  pos4spacer,
  pos5name,
  pos5spacer,
  DR1 = "AATTTCTACTGTCGTAGAT",
  DR16 = "AATTCCTACTATTGTAGGT",
  DR10 = "AATTCCTACTCTCGTAGGT",
  DR3 = "AATTTCTACTCTAGTAGAT",
  Gibson5p = "atcttgtggaaaggacgaaacaccgaatttctactcttgt",
  Gibson3p = "AATTTCTCCTCTAGGAGATtttttttaagcttggcgtaactagatcttga",
  outputdir = getwd()
)
}
\value{
list of data frames containing 1. oligos (this is too expensive but is kept as an available alternative for now) and 2. gene block
}
\description{
for individually cloning Cas12a 5-plex spacer array into vector with BsmBI site (e.g. pRG212). Modified from protocol by Qingzhou Chen and Junwei Shi. Variant DR sequences from Deweirdt et al. are used to minimize recombination. Also as much as possible avoid TTTCT motif (weak Pol III terminator)
Returns a list with two data frames, one containing oligos/ultramers, the other containing gene block. User can choose to use either one for synthesis, but gene block is more cost effective.
For gene block approach, by default insert is appended with homology regions (Gibson5p and Gibson3p as input arguments) compatible with Gibson assembly into pCH49.
For oligos/ultramer approach: nomenclature for naming sense and antisense strands is as follows: sense1 anneals with antisense1, sense2 anneals with antisense2, sense 3 anneals with antisense3, etc. These prefixes are printed at the front of the string to facilitate setting up annealing reactions because the IDT tubes truncates the end of the long string.
by Chris Hsiung, updated 2021-12-23
}
