% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOligosCas12a6plexgRNAVariantDR_BsmbI.R
\name{getOligosCas12a6plexgRNAVariantDR_BsmbI}
\alias{getOligosCas12a6plexgRNAVariantDR_BsmbI}
\title{for individually cloning Cas12a 6-plex spacer array into vector with BsmBI site (e.g. pRG212). Modified from protocol by Qingzhou Chen and Junwei Shi. Variant DR sequences from Deweirdt et al. are used to minimize recombination.
Nomenclature for naming sense and antisense strands is as follows: sense1 anneals with antisense1, sense2 anneals with antisense2, sense 3 anneals with antisense3, etc. These prefixes are printed at the front of the string to facilitate setting up annealing reactions because the IDT tubes truncates the end of the long string.
Returns a list with ultramers OR 60-mer style oligos}
\usage{
getOligosCas12a6plexgRNAVariantDR_BsmbI(
  pos1name,
  pos1spacer,
  pos2name,
  pos2spacer,
  pos3name,
  pos3spacer,
  pos4name,
  pos4spacer,
  pos5name,
  pos5spacer,
  pos6name,
  pos6spacer,
  DRWT = "aatttctactcttgtagat",
  DR1 = "TAATTTCTACTGTCGTAGAT",
  DR2 = "TAATTTCTACTATCGTAGAT",
  DR3 = "AAATTTCTACTCTAGTAGAT",
  DR5 = "TaatttctactAtAgtagat",
  DR6 = "AaatttctCctctCgGagat",
  outputdir = getwd()
)
}
\value{
data frame of oligos
}
\description{
for individually cloning Cas12a 6-plex spacer array into vector with BsmBI site (e.g. pRG212). Modified from protocol by Qingzhou Chen and Junwei Shi. Variant DR sequences from Deweirdt et al. are used to minimize recombination.
Nomenclature for naming sense and antisense strands is as follows: sense1 anneals with antisense1, sense2 anneals with antisense2, sense 3 anneals with antisense3, etc. These prefixes are printed at the front of the string to facilitate setting up annealing reactions because the IDT tubes truncates the end of the long string.
Returns a list with ultramers OR 60-mer style oligos
}
