% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOligosCas12aTriplegRNA_BsmbI.R
\name{getOligosCas12aTriplegRNA_BsmbI}
\alias{getOligosCas12aTriplegRNA_BsmbI}
\title{for individually cloning Cas12a triple spacer array into vector withi BsmBI site (e.g. pRG212)}
\usage{
getOligosCas12aTriplegRNA_BsmbI(
  pos1name,
  pos1spacer,
  pos2name,
  pos2spacer,
  pos3name,
  pos3spacer,
  DR = "aatttctactcttgtagat",
  outputdir = getwd()
)
}
\description{
for individually cloning Cas12a triple spacer array into vector withi BsmBI site (e.g. pRG212)
}
