% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOligosCas12aTriplegRNAVariantAsDR_BsmbI.R
\name{getOligosCas12aTriplegRNAVariantAsDR_BsmbI}
\alias{getOligosCas12aTriplegRNAVariantAsDR_BsmbI}
\title{for individually cloning Cas12a triple spacer array into vector with BsmBI site containing a 3' DR in the style of pRG212 (see Gier et al. 2020, uses As DR WT on 5' end and As DR WT on 3' end). For middle 2 spacers, uses variant As DR1 and DR3 sequences from DeWeirdt et al., excluding the first nt in Fig. 4c, i.e. each DR is 19nt, same as in pRG212).  Also compatible with pCH28 (uses As DR WT on 5' end and As DR8 variant on 3' end).
Nomenclature for naming sense and antisense strands is as follows: sense1 anneals with antisense1, sense2 anneals with antisense2, sense 3 anneals with antisense3, etc. These prefixes are printed at the front of the string to facilitate setting up annealing reactions because the IDT tubes truncates the end of the long string.
updated 9/9/21, updated 2023-05-15}
\usage{
getOligosCas12aTriplegRNAVariantAsDR_BsmbI(
  pos1name,
  pos1spacer,
  pos2name,
  pos2spacer,
  pos3name,
  pos3spacer,
  AsDR1 = "AATTTCTACTGTCGTAGAT",
  AsDR3 = "AATTTCTACTCTAGTAGAT",
  outputdir = getwd()
)
}
\value{
data frame of oligos
}
\description{
for individually cloning Cas12a triple spacer array into vector with BsmBI site containing a 3' DR in the style of pRG212 (see Gier et al. 2020, uses As DR WT on 5' end and As DR WT on 3' end). For middle 2 spacers, uses variant As DR1 and DR3 sequences from DeWeirdt et al., excluding the first nt in Fig. 4c, i.e. each DR is 19nt, same as in pRG212).  Also compatible with pCH28 (uses As DR WT on 5' end and As DR8 variant on 3' end).
Nomenclature for naming sense and antisense strands is as follows: sense1 anneals with antisense1, sense2 anneals with antisense2, sense 3 anneals with antisense3, etc. These prefixes are printed at the front of the string to facilitate setting up annealing reactions because the IDT tubes truncates the end of the long string.
updated 9/9/21, updated 2023-05-15
}
