% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReverseComplement.R
\name{getReverseComplement}
\alias{getReverseComplement}
\title{Wrapper for Biostring::reverseComplement()}
\usage{
getReverseComplement(sequence)
}
\description{
Wrapper for Biostring::reverseComplement()
}
