% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpacersforGeneByRank.R
\name{getSpacersForGeneByRank}
\alias{getSpacersForGeneByRank}
\title{for given gene retrieve top rows corresponding to (number of rows specified by numbertoget) after sorting by any numerical score in rankcol
libdf can be any data frame (each row is a sgRNA protospacer sequence) that contains columns specified by the other input arguments}
\usage{
getSpacersForGeneByRank(
  gene,
  libdf,
  numbertoget,
  rankcol = "selection rank",
  genecol = "gene",
  spacercol = "protospacer sequence"
)
}
\value{
dataframe with picked rows with additional pickrank column to indicate rank after sorting by rankcol
}
\description{
for given gene retrieve top rows corresponding to (number of rows specified by numbertoget) after sorting by any numerical score in rankcol
libdf can be any data frame (each row is a sgRNA protospacer sequence) that contains columns specified by the other input arguments
}
