% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsplitviolin_withthres.R
\name{plotsplitviolin_withthres}
\alias{plotsplitviolin_withthres}
\title{Plot Split Violin with Thresholds}
\usage{
plotsplitviolin_withthres(
  data,
  x_var,
  y_var,
  fill_var,
  dodge_width = 0.8,
  color_palette = c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00",
    "#CC79A7"),
  facetformula_var,
  title_var,
  thres_var,
  perc_var,
  perclabely = 2,
  labelsize = 6,
  basesize_var = 18
)
}
\arguments{
\item{data}{A data frame containing the variables used for plotting.}

\item{x_var}{The variable name for the x-axis.}

\item{y_var}{The variable name for the y-axis.}

\item{fill_var}{The variable name for the fill color.}

\item{dodge_width}{The width of the dodge for grouped violins (default = 0.8).}

\item{color_palette}{A vector of colors to use for the plot (default = cbbPalette).}

\item{facetformula_var}{The formula specifying the faceting variables.}

\item{title_var}{The title for the plot.}

\item{thres_var}{The variable name for the threshold.}

\item{perc_var}{The variable name for the percentage label.}

\item{labelsize}{The font size for the percentage label (default = 6).}

\item{basesize_var}{The base font size for the plot (default = 18).}

\item{facet_strvar}{The variable name for facet labels.}

\item{perclabelx}{The x-coordinate adjustment for the percentage label (default = 1).}

\item{facet_var}{The variable name for faceting (optional).}
}
\value{
A ggplot object representing the half violin plot with thresholds.
}
\description{
note for this to work properly, subset data frame ahead of time such that there fill_var is a factor of 2 levels -CH 2023-09-02
}
\details{
This function creates a split violin plot with thresholds using ggplot2.
}
