% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proptestfordf.R
\name{proptestfordf}
\alias{proptestfordf}
\title{Proportion Test for Two Groups in a DataFrame}
\usage{
proptestfordf(
  data,
  response_col,
  group_col,
  group1_label,
  group2_label,
  alternative_var = "two-sided"
)
}
\arguments{
\item{data}{A dataframe containing the data to be tested.}

\item{response_col}{The name of the column in the dataframe that contains the binary response variable.}

\item{group_col}{The name of the column in the dataframe that specifies the group.}

\item{group1_label}{The label for the first group to be tested.}

\item{group2_label}{The label for the second group to be tested.}
}
\value{
A dataframe containing the labels for the two groups and the p-value from the proportion test.
}
\description{
This function performs a proportion test between two specified groups in a given dataframe.
It calculates the number of successes for each group and then conducts a proportion test.
}
\examples{
\dontrun{
df <- data.frame(
  response = c(1, 0, 1, 0, 1, 0, 1, 0),
  group = c('A', 'A', 'B', 'B', 'A', 'A', 'B', 'B')
)
result <- proptestfordf(df, 'response', 'group', 'A', 'B')
print(result)
}

2024-03-23 added alternative_var
}
