% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAfold2df.R
\name{RNAfold2df}
\alias{RNAfold2df}
\title{Function to parse the output from RNAfold 2.4.18 into data frame and write to tab-delimited file
parse output according to https://www.tbi.univie.ac.at/RNA/RNAfold.1.html#heading5
this description is based on manual entry of a single sequence -- "Here, the first line just repeats the sequence input. The second line contains a MFE structure in dot bracket notation followed by the minimum free energy. After this, the pairing probabilities for each nucleotide are shown in a pseudo dot-bracket notation followed by the free energy of ensemble. The next two lines show the centroid structure with its free energy and its distance to the ensemble as well as the MEA structure, its free energy and the maximum expected accuracy, respectively. The last line finally contains the frequency of the MFE representative in the complete ensemble of secondary structures and the ensemble diversity."
the output for multiple sequences contains same info formatted slightly differently, figured it out and parsed accordingly.
writes a tab-delimited file with the in the format '{basename of input file}df.txt'}
\usage{
RNAfold2df(file, writedir = getwd())
}
\description{
Function to parse the output from RNAfold 2.4.18 into data frame and write to tab-delimited file
parse output according to https://www.tbi.univie.ac.at/RNA/RNAfold.1.html#heading5
this description is based on manual entry of a single sequence -- "Here, the first line just repeats the sequence input. The second line contains a MFE structure in dot bracket notation followed by the minimum free energy. After this, the pairing probabilities for each nucleotide are shown in a pseudo dot-bracket notation followed by the free energy of ensemble. The next two lines show the centroid structure with its free energy and its distance to the ensemble as well as the MEA structure, its free energy and the maximum expected accuracy, respectively. The last line finally contains the frequency of the MFE representative in the complete ensemble of secondary structures and the ensemble diversity."
the output for multiple sequences contains same info formatted slightly differently, figured it out and parsed accordingly.
writes a tab-delimited file with the in the format '{basename of input file}df.txt'
}
