% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_groupeddf.R
\name{subsample_groupeddf}
\alias{subsample_groupeddf}
\title{Subsample a Data Frame}
\usage{
subsample_groupeddf(df, Nthres)
}
\arguments{
\item{df}{A data frame to be subsampled. Can be grouped or ungrouped.}

\item{Nthres}{An integer threshold for subsampling. If a group in \code{df} (or the entire \code{df} if it's not grouped)
has fewer rows than \code{Nthres}, all rows of that group (or the entire data frame) are returned.
Otherwise, a random subset of rows, of size \code{Nthres}, is returned.}
}
\value{
A subsampled data frame.
}
\description{
The \code{subsample_groupeddf} function returns a subsampled version of the input data frame. If the number of rows
in a group of the input data frame is less than the specified threshold (\code{Nthres}), all rows of that
group are returned. Otherwise, a random subset of rows, of size \code{Nthres}, from each group is returned.
}
\examples{
\dontrun{
data <- data.frame(group = rep(1:2, each = 100), value = rnorm(200))
subsampled_data <- subsample_groupeddf(data, 50)
}

}
