% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxtestfordf.R
\name{wilcoxtestfordf}
\alias{wilcoxtestfordf}
\title{Wilcoxon Test for Two Groups in a DataFrame}
\usage{
wilcoxtestfordf(
  dataframe,
  groupvar1,
  groupvar2,
  groupname_var,
  value_var,
  alternative_var,
  round_var = 2,
  sciformat_var = FALSE
)
}
\arguments{
\item{dataframe}{A data frame containing the data.}

\item{groupvar1}{The first group to be compared.}

\item{groupvar2}{The second group to be compared.}

\item{groupname_var}{The column name in dataframe that represents the grouping variable.}

\item{value_var}{The column name in dataframe representing the values on which the Wilcoxon test is performed.}

\item{alternative_var}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}
}
\value{
A data frame with the p-value of the Wilcoxon test. If there's insufficient data for either group, the result will be "Insufficient data".
}
\description{
Performs a Wilcoxon test on two groups within a dataframe based on a grouping variable and returns the p-value.
}
\examples{
\dontrun{
df <- data.frame(group = c("A", "A", "B", "B"), value = c(1,2,3,4))
wilcoxtestfordf(df, "A", "B", "group", "value", "less")
}

}
