% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writedf2fasta.R
\name{writedf2fasta}
\alias{writedf2fasta}
\title{wrapper for seqinr::write.fasta that takes data frame as input}
\usage{
writedf2fasta(df, seqvar, namevar, fileout)
}
\value{
writes FASTA file
}
\description{
wrapper for seqinr::write.fasta that takes data frame as input
}
