% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLibraryOligoCas12aTriplegRNA.R
\name{getLibraryOligoCas12aTriplegRNA}
\alias{getLibraryOligoCas12aTriplegRNA}
\title{Function to assemble 6-plex oligo for pooled synthesis for Cas12a 6-plex gRNA expression
The oligo will be PCR'd using adaptor sequences, then digested with BsmbI and ligated into backbone (e.g. compatible with pCH49)
by Chris Hsiung 2022-03-03}
\usage{
getLibraryOligoCas12aTriplegRNA(
  pos1name,
  pos1spacer,
  pos2name,
  pos2spacer,
  pos3name,
  pos3spacer,
  Adaptor5p,
  Adaptor3p,
  DR1 = "AATTTCTACTGTCGTAGAT",
  DR3 = "AATTTCTACTCTAGTAGAT",
  BsmbI5p = "cgtctcAAGAT",
  BsmbI3p = "AATTcgagacg",
  outputdir = getwd()
)
}
\value{
A list containing oligoname and oligosequence
}
\description{
Function to assemble 6-plex oligo for pooled synthesis for Cas12a 6-plex gRNA expression
The oligo will be PCR'd using adaptor sequences, then digested with BsmbI and ligated into backbone (e.g. compatible with pCH49)
by Chris Hsiung 2022-03-03
}
