% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOligosCas12aDualgRNA_BsmbI.R
\name{getOligosCas12aDualgRNA_BsmbI}
\alias{getOligosCas12aDualgRNA_BsmbI}
\title{not thoroughly vetted yet}
\usage{
getOligosCas12aDualgRNA_BsmbI(
  pos1name,
  pos1spacer,
  pos2name,
  pos2spacer,
  outputdir = getwd()
)
}
\description{
not thoroughly vetted yet
}
