% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFlowjoExportDir.R
\name{processFlowjoExportDir}
\alias{processFlowjoExportDir}
\title{Function to batch process .csv files (containing the substring specified by 'csvstring') exported from flowjo within the directory specified by 'dir'. Directory must contain 'platesetupfile' (a tab-delimited .txt file) that contains column with name specified by 'IDvar' (defaults to 'Sample') that should contain 96well ID's that are matched to the $FIL keyword in the .fcs file exported from the Attune.
Outputs a tab-delimited .txt file with file name specified by 'fileout'.}
\usage{
processFlowjoExportDir(
  dir,
  csvstring = "_P1.csv",
  fileout = "P1_combodf.txt",
  platesetupfile = "platesetup.txt",
  IDvar = "Sample"
)
}
\description{
Function to batch process .csv files (containing the substring specified by 'csvstring') exported from flowjo within the directory specified by 'dir'. Directory must contain 'platesetupfile' (a tab-delimited .txt file) that contains column with name specified by 'IDvar' (defaults to 'Sample') that should contain 96well ID's that are matched to the $FIL keyword in the .fcs file exported from the Attune.
Outputs a tab-delimited .txt file with file name specified by 'fileout'.
}
